# specs/001-warehouse-inventory-system/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse Inventory Management API
  description: |
    REST API per la gestione del magazzino officina con registrazione movimenti (IN/OUT/ADJUSTMENT),
    inventario in tempo reale, e funzionalità di export Excel.

    ## Architettura
    - **Backend**: FastAPI con validazione Pydantic
    - **Database**: PostgreSQL 15+ con event-sourcing (movimenti immutabili)
    - **Frontend**: Vercel (Next.js)

    ## Caratteristiche Principali
    - Giacenze calcolate in tempo reale da movimenti
    - Validazione movimenti OUT per prevenire giacenze negative
    - Rettifiche inventario (ADJUSTMENT) con nota obbligatoria
    - Export Excel lato client (ultimi 12 mesi)
    - Supporto per quantità decimali (peso/volume)

    ## Localizzazione
    Tutti i messaggi di errore e le descrizioni sono in italiano.
  version: 1.0.0
  contact:
    name: Warehouse Inventory System
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server (localhost)
  - url: https://inventario-magazzino.onrender.com
    description: Production server (Render)
...
paths:
...
  /api/movements:
    get:
      tags: [Movements]
      summary: Lista movimenti con filtri
      description: |
        Restituisce lo storico dei movimenti di magazzino con opzioni di filtro per data,
        articolo, e tipo di movimento. Default: ultimi 30 giorni.

        **Requisiti**: FR-022, FR-023, FR-024
      operationId: listMovements
      parameters:
        - name: from_date
          in: query
          description: Data inizio range (inclusive)
          required: false
          schema:
            type: string
            format: date
            example: "2025-10-01"
        - name: to_date
          in: query
          description: Data fine range (inclusive)
          required: false
          schema:
            type: string
            format: date
            example: "2025-11-11"
        - name: item_id
          in: query
          description: Filtra per articolo specifico
          required: false
          schema:
            type: string
            format: uuid
        - name: movement_type
          in: query
          description: Filtra per tipo di movimento
          required: false
          schema:
            type: string
            enum: [IN, OUT, ADJUSTMENT]
        - name: limit
          in: query
          description: Numero massimo di risultati
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          description: Numero di record da saltare (paginazione)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Lista movimenti
          content:
            application/json:
              schema:
                type: object
                properties:
                  movements:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovementDetail'
                  total:
                    type: integer
                    description: Numero totale di movimenti (senza paginazione)
                  limit:
                    type: integer
                  offset:
                    type: integer
              examples:
                success:
                  summary: Storico movimenti
                  value:
                    movements:
                      - id: "m1a2b3c4-d5e6-7890-abcd-ef1234567890"
                        item_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                        item_name: "Olio motore 5W30"
                        movement_type: "IN"
                        quantity: 20.0
                        unit: "lt"
                        movement_date: "2025-11-10"
                        timestamp: "2025-11-10T14:30:00Z"
                        unit_cost_override: 8.50
                        note: "Carico fornitore A - Fattura 456"
                      - id: "m2b3c4d5-e6f7-8901-bcde-f12345678901"
                        item_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                        item_name: "Filtro olio"
                        movement_type: "OUT"
                        quantity: -3.0
                        unit: "pz"
                        movement_date: "2025-11-08"
                        timestamp: "2025-11-08T09:15:00Z"
                        unit_cost_override: null
                        note: "Tagliando Alfa Romeo 159"
                      - id: "m3c4d5e6-f7a8-9012-cdef-123456789012"
                        item_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                        item_name: "Filtro olio"
                        movement_type: "ADJUSTMENT"
                        quantity: -2.0
                        unit: "pz"
                        movement_date: "2025-11-05"
                        timestamp: "2025-11-05T16:45:00Z"
                        unit_cost_override: null
                        note: "Conteggio fisico mensile - 2 unità danneggiate"
                    total: 128
                    limit: 100
                    offset: 0
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [Movements]
      summary: Registra movimento
      description: |
        Registra un nuovo movimento di magazzino (IN/OUT/ADJUSTMENT).
        Ogni tipo ha validazioni specifiche:

        - **IN**: quantità > 0, optional unit_cost_override, optional note
        - **OUT**: quantità > 0 (convertita in negativo), validazione giacenza disponibile, conferma richiesta, optional note
        - **ADJUSTMENT**: target_stock (giacenza desiderata), nota obbligatoria, calcolo automatico delta

        **Requisiti**: FR-006, FR-007, FR-008, FR-009, FR-010, FR-011, FR-012, FR-036, FR-037, FR-038, FR-039
      operationId: createMovement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/MovementInCreate'
                - $ref: '#/components/schemas/MovementOutCreate'
                - $ref: '#/components/schemas/MovementAdjustmentCreate'
              discriminator:
                propertyName: movement_type
                mapping:
                  IN: '#/components/schemas/MovementInCreate'
                  OUT: '#/components/schemas/MovementOutCreate'
                  ADJUSTMENT: '#/components/schemas/MovementAdjustmentCreate'
            examples:
              carico:
                summary: Carico merce (IN)
                value:
                  movement_type: "IN"
                  item_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  quantity: 25.0
                  movement_date: "2025-11-11"
                  unit_cost_override: 8.75
                  note: "Carico fornitore B - Fattura 789"
              scarico:
                summary: Scarico merce (OUT)
                value:
                  movement_type: "OUT"
                  item_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                  quantity: 2.0
                  movement_date: "2025-11-11"
                  note: "Usato per riparazione Fiat Panda"
                  confirmed: true
              rettifica_positiva:
                summary: Rettifica in aumento
                value:
                  movement_type: "ADJUSTMENT"
                  item_id: "c3d4e5f6-a7b8-9012-cdef-123456789012"
                  target_stock: 15.0
                  movement_date: "2025-11-11"
                  note: "Conteggio fisico: trovate 5 unità in magazzino secondario"
              rettifica_negativa:
                summary: Rettifica in diminuzione
                value:
                  movement_type: "ADJUSTMENT"
                  item_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                  target_stock: 8.0
                  movement_date: "2025-11-11"
                  note: "Conteggio fisico mensile - 3 unità danneggiate da infiltrazione acqua"
      responses:
        '201':
          description: Movimento registrato con successo
          headers:
            Location:
              description: URI del nuovo movimento
              schema:
                type: string
                example: "/api/movements/m1a2b3c4-d5e6-7890-abcd-ef1234567890"
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementDetail'
              examples:
                in_success:
                  summary: Carico registrato
                  value:
                    id: "m1a2b3c4-d5e6-7890-abcd-ef1234567890"
                    item_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    item_name: "Olio motore 5W30"
                    movement_type: "IN"
                    quantity: 25.0
                    unit: "lt"
                    movement_date: "2025-11-11"
                    timestamp: "2025-11-11T10:30:45Z"
                    unit_cost_override: 8.75
                    note: "Carico fornitore B - Fattura 789"
                out_success:
                  summary: Scarico registrato
                  value:
                    id: "m2b3c4d5-e6f7-8901-bcde-f12345678901"
                    item_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                    item_name: "Filtro olio"
                    movement_type: "OUT"
                    quantity: -2.0
                    unit: "pz"
                    movement_date: "2025-11-11"
                    timestamp: "2025-11-11T10:35:22Z"
                    unit_cost_override: null
                    note: "Usato per riparazione Fiat Panda"
                adjustment_success:
                  summary: Rettifica registrata
                  value:
                    id: "m3c4d5e6-f7a8-9012-cdef-123456789012"
                    item_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                    item_name: "Filtro olio"
                    movement_type: "ADJUSTMENT"
                    quantity: -3.0
                    unit: "pz"
                    movement_date: "2025-11-11"
                    timestamp: "2025-11-11T10:40:15Z"
                    unit_cost_override: null
                    note: "Conteggio fisico mensile - 3 unità danneggiate da infiltrazione acqua"
        '400':
          description: Validazione fallita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_quantity:
                  summary: Quantità non valida
                  value:
                    detail: "La quantità deve essere maggiore di zero"
                    error_code: "INVALID_QUANTITY"
                missing_note_adjustment:
                  summary: Nota mancante per rettifica
                  value:
                    detail: "La nota è obbligatoria per le rettifiche (spiega il motivo della discrepanza)"
                    error_code: "ADJUSTMENT_NOTE_REQUIRED"
                no_adjustment_needed:
                  summary: Nessuna rettifica necessaria
                  value:
                    detail: "La giacenza target (15.0) coincide con quella attuale. Nessuna rettifica necessaria."
                    error_code: "NO_ADJUSTMENT_NEEDED"
                invalid_date:
                  summary: Data non valida
                  value:
                    detail: "La data del movimento non può essere nel futuro o più di 365 giorni nel passato"
                    error_code: "INVALID_MOVEMENT_DATE"
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflitto - Giacenza insufficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_stock:
                  summary: Giacenza insufficiente per scarico
                  value:
                    detail: "Impossibile scaricare 5.0 pz. Giacenza disponibile: 3.0 pz"
                    error_code: "INSUFFICIENT_STOCK"
                    context:
                      requested_quantity: 5.0
                      available_stock: 3.0
                      unit: "pz"
  /api/movements/{id}:
    get:
      tags: [Movements]
      summary: Dettaglio movimento
      description: Restituisce i dettagli completi di un movimento specifico
      operationId: getMovement
      parameters:
        - name: id
          in: path
          description: ID del movimento
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Movimento trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    MovementInCreate:
      type: object
      required:
        - movement_type
        - item_id
        - quantity
      properties:
        movement_type:
          type: string
          enum: [IN]
          description: Tipo movimento (carico)
          example: "IN"
        item_id:
          type: string
          format: uuid
          description: ID articolo
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        quantity:
          type: number
          format: decimal
          minimum: 0.001
          exclusiveMinimum: true
          description: Quantità in ingresso (deve essere > 0)
          example: 25.0
        movement_date:
          type: string
          format: date
          description: Data movimento (default oggi, può essere passato recente)
          example: "2025-11-11"
        unit_cost_override:
          type: number
          format: decimal
          minimum: 0
          nullable: true
          description: Override costo unitario (aggiorna anche articolo.unit_cost)
          example: 8.75
        note:
          type: string
          nullable: true
          description: Nota opzionale
          example: "Carico fornitore B - Fattura 789"
    MovementOutCreate:
      type: object
      required:
        - movement_type
        - item_id
        - quantity
        - confirmed
      properties:
        movement_type:
          type: string
          enum: [OUT]
          description: Tipo movimento (scarico)
          example: "OUT"
        item_id:
          type: string
          format: uuid
          description: ID articolo
          example: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        quantity:
          type: number
          format: decimal
          minimum: 0.001
          exclusiveMinimum: true
          description: Quantità in uscita (inserita come positivo, salvata come negativo)
          example: 3.0
        movement_date:
          type: string
          format: date
          description: Data movimento
          example: "2025-11-11"
        note:
          type: string
          nullable: true
          description: Nota opzionale
          example: "Usato per riparazione Fiat Panda"
        confirmed:
          type: boolean
          description: Conferma esplicita richiesta per scarico (FR-009)
          example: true
    MovementAdjustmentCreate:
      type: object
      required:
        - movement_type
        - item_id
        - target_stock
        - note
      properties:
        movement_type:
          type: string
          enum: [ADJUSTMENT]
          description: Tipo movimento (rettifica)
          example: "ADJUSTMENT"
        item_id:
          type: string
          format: uuid
          description: ID articolo
          example: "c3d4e5f6-a7b8-9012-cdef-123456789012"
        target_stock:
          type: number
          format: decimal
          minimum: 0
          description: Giacenza target desiderata (sistema calcola delta automaticamente)
          example: 15.0
        movement_date:
          type: string
          format: date
          description: Data rettifica
          example: "2025-11-11"
        note:
          type: string
          minLength: 1
          description: Nota OBBLIGATORIA che spiega motivo della discrepanza
          example: "Conteggio fisico mensile - trovate 5 unità in magazzino secondario"
    MovementDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificativo movimento
          example: "m1a2b3c4-d5e6-7890-abcd-ef1234567890"
        item_id:
          type: string
          format: uuid
          description: ID articolo
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        item_name:
          type: string
          description: Nome articolo (denormalizzato per visualizzazione)
          example: "Olio motore 5W30"
        movement_type:
          type: string
          enum: [IN, OUT, ADJUSTMENT]
          description: Tipo movimento
          example: "IN"
        quantity:
          type: number
          format: decimal
          description: Quantità movimento (IN positivo, OUT negativo, ADJUSTMENT +/-)
          example: 25.0
        unit:
          type: string
          description: Unità di misura (denormalizzato)
          example: "lt"
        movement_date:
          type: string
          format: date
          description: Data movimento
          example: "2025-11-11"
        timestamp:
          type: string
          format: date-time
          description: Timestamp server-side di creazione
          example: "2025-11-11T10:30:45Z"
        unit_cost_override:
          type: number
          format: decimal
          nullable: true
          description: Costo unitario override (solo per IN)
          example: 8.75
        note:
          type: string
          nullable: true
          description: Nota (obbligatoria per ADJUSTMENT)
          example: "Carico fornitore B - Fattura 789"
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Messaggio errore in italiano
          example: "Impossibile scaricare 5.0 pz. Giacenza disponibile: 3.0 pz"
        error_code:
          type: string
          description: Codice errore machine-readable
          example: "INSUFFICIENT_STOCK"
        context:
          type: object
          nullable: true
          description: Informazioni aggiuntive specifiche per l'errore
          additionalProperties: true
          example:
            requested_quantity: 5.0
            available_stock: 3.0
            unit: "pz"
...
  responses:
    BadRequest:
      description: Richiesta non valida (validazione Pydantic fallita)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Errore validazione
              value:
                detail: "Errore di validazione: il campo 'name' è obbligatorio"
                error_code: "VALIDATION_ERROR"
                context:
                  field: "name"
                  constraint: "required"
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            item_not_found:
              summary: Articolo non trovato
              value:
                detail: "Articolo con ID 'a1b2c3d4-e5f6-7890-abcd-ef1234567890' non trovato"
                error_code: "ITEM_NOT_FOUND"
            movement_not_found:
              summary: Movimento non trovato
              value:
                detail: "Movimento con ID 'm1a2b3c4-d5e6-7890-abcd-ef1234567890' non trovato"
                error_code: "MOVEMENT_NOT_FOUND"
  securitySchemes:
    # Nessuna autenticazione per MVP (single-user)
    # Placeholder per future implementazioni (JWT, API Key, etc.)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticazione JWT (non implementata in MVP).
        Per uso futuro con multi-utenza.

# Security: Nessuna autenticazione richiesta per MVP (single-user, accesso diretto)
security: []